refactor: add unified reader utilities
